+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 0.5.0
 -----
 
+- Add :ref:`block_level_ignore` and :ref:`page_level_ignore`
+  `#33 <https://github.com/jdillard/sphinx-llms-txt/pull/33>`_
 - Add :confval:`llms_txt_full_size_policy` configuration option to control behavior when :confval:`llms_txt_full_max_size` is exceeded.
   `#35 <https://github.com/jdillard/sphinx-llms-txt/pull/35>`_